backend/requirements/base.txt
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
drf-jwt==1.19.1
psycopg2==2.9.1
Faker==8.11.0
Twisted==21.7.0
gunicorn==20.1.0
uvicorn[standard]==0.15.0
django-mjml==0.11.0
requests==2.26.0
itsdangerous==2.0.1
drf-spectacular==0.13.1
Pillow==8.3.1
asgiref==3.4.1
channels==3.0.4
channels-redis==3.3.0
celery[redis]==5.1.2
django-celery-email==3.0.0
advocate==1.0.0
zipp==3.5.0
unicodecsv==0.14.1
django-celery-beat==2.2.1
celery-redbeat==2.0.0
service-identity==21.1.0
regex==2021.8.3
click==7.1.2
cryptography==3.4.8
antlr4-python3-runtime==4.8.0
tqdm==4.62.3
